refactor(privateMode): move option to general tab and requires refresh
